[io] Added low level API for file download
  * Moved sendfile code to the Local File System Storage
  * Issue: BNN-279
